third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-host-expected.txt

Tests ShadowRoot.host property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS root.host is host
PASS root.host is host
PASS roots[0].host is host
PASS roots[1].host is host
PASS roots[0].host is host
PASS roots[1].host is host
PASS roots[1].host is host
PASS roots[6].host is host
PASS root.host is host




PASS successfullyParsed is true

TEST COMPLETE
